moving files for project structure
